+++ advertion/public.py
@@ -9,7 +9,7 @@ from .core import AdversarialValidation
 def validate(
     trainset: pd.DataFrame,
     testset: pd.DataFrame,
-    target: str,
+    target: str = None,
     smart: bool = True,
     n_splits: int = 5,
     verbose: bool = True,
@@ -20,12 +20,12 @@ def validate(
     Args:
         trainset (pd.DataFrame): The training dataset.
         testset (pd.DataFrame): The test dataset.
-        target (str): The target column name.
-        smart (bool, optional): Whether to prune features with strongly identifiable properties.
-        n_splits (int, optional): The number of splits to perform.
-        verbose (bool, optional): Whether to print informative messages to the standard output.
+        target (str): The target column name. Default is None.
+        smart (bool, optional): Whether to prune features with strongly identifiable properties. Default is True.
+        n_splits (int, optional): The number of splits to perform. Default is 5.
+        verbose (bool, optional): Whether to print informative messages to the standard output. Default is True.
         random_state (Union[int, np.random.RandomState], optional): If you wish to ensure reproducible output across \
-        multiple function calls.
+        multiple function calls. Default is None.
 
     Returns:
         dict: An informative key-valued response.
@@ -42,7 +42,6 @@ def validate(
         >>> validate(
         >>>     trainset=train,
         >>>     testset=test,
-        >>>     target="label",
         >>> )
 
         >>> // {